Fewer temptables for new snapshot
We allocate one temptable per log file for pglogs. When the database
accumulates a large number of log files (a long running snapshot transaction,
node crashed, etc.), many temptables will be created and it may eventually
cause the system to run out of memory.

This patch reduces the number of temptables needed: only a fixed number of
temptables are created, and are shared among log files in a round-robin fashion.

(DRQS 168188318)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>