Fix upsert edge case in delayed_key_add
When adding an index, if the sc pointer is on the right of an upsert (in other words, the stripe has finished converting before the upsert lands on the stripe),
we may either skip adding the new record to the new index (which results in a missing key verify failure), or even worse add wrong data to the new index.

This patch fixes it.

(DRQS 168963987)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>