Fix OP_Rewind crash
This is cherry picked from https://www.sqlite.org/src/info/778b1224a318d013.
Without this change we might mistakenly code an OP_Gosub to a cancelled subroutine.
A cancelled subroutine has no return address which means that we would not be able
to come back from the subroutine to the OP_Gosub opcode. This would lead to incorrect
execution of the opcodes.

(DRQS 169753545)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>