Turn unknown files to DB_DELETED on recovery for dbs with ufid enabled.
Can't get a consistent test case for this. But enough schema changes followed
by files going away followed by database bounce will cause us to not be
able to find files mentioned in logs during recovery. This was handled
correctly for dbreg, but not ufid. This applies the same change for ufid.

Signed-off-by: Mike Ponomarenko <mponomarenko@bloomberg.net>
Signed-off-by: Akshat Sikarwar <asikarwar1@bloomberg.net>