Fix sc throttle edge case
A convert thread throttles itself when the number of log bytes that have been generated so far is above a threshold.
However multiple convert threads can check the current number of log bytes at the same time, all thinking
that they need not throttle. The total number of log bytes from all these convert threads, as a result,
can be way over the threshold (rebuilding lengthy blob records, for instance).

This patch attempts to fix it. We estimate how much future log traffic an sc tran will generate, and throttle
if the number is above the threshold.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>