Tweak the delayed-file-deletion logic
Make sure that we use an updated most recent log file every time we queue a file for deletion.
Otherwise on a busy database we may wind up deleting files too aggressively. The pseudo code
below explains the bug:

```
N = current-log-file
for file in db-dir
    if file is unused
        /*
         * BUG: A log file younger than N (let's say, N+5) could have
         * produced this btree (e.g., a lot writes followed by a fastinit).
         * And we end up using an outdated log file for this btree.
         */
        add {fname:file, lognum:N} to old-file-list
    end
end

for file in old-file-list
    if file.lognum < lowest-lognum-cluster-wide
        /*
         * The file is deleted here while it should've been kept around
         * till lowest-lognum-cluster-wide is past N+5
         */
        remove file.fname
    end
end
```

To fix this, we simply grab the most recent log file before we add a file to the old-file list, as below

```diff
-- N = current-log-file
for file in db-dir
    if file is unused
++      N = current-log-file
        add {fname:file, lognum:N} to old-file-list
    end
end
```

(DRQS 170280654)

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>