Improving log->put performance
This patch introduces a new tunable to toggle the O_SYNC flag for log files.
If O_SYNC isn't used, after the log buffer is written out when it's full,
an synchronous sync_file_range() call is issued on the log file, to speed up future
fsync() calls. In other words, We turn the synchronous IO in log->put() into an
asynchronous write-out of page cache, and hence we can move much of the IO latency
off the critical path. log->flush(), however, does issue an fsync() call after
flushing out the log buffer, to ensure data has made it to disk.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>