Fix an issue with comdb2_sc_status
In version R8, some backwards incompatible changes got introduced into
the schema change object that broke the object's original deserializer
function (buf_get_schemachange()). As a result, reading an sc status object
created by R7 would fail if read by R8 (via comdb2_sc_status).

The fix was to keep the both the versions of the deserializer functions and
invoke them appropriately.

The current (potential hackish) method to pick the right version on the
deserializer function is based on the content of the first 4 bytes of the
LLMETA_SCHEMACHANGE_STATUS payload, where it is assumed that the valid
values of s->kind (between SC_INVALID and SC_LAST, exclusive) will not
coincide with the first 4 bytes of the rqid (fastseed) stored as the first
member in old (7.0's) LLMETA_SCHEMACHANGE_STATUS payload.

Refer: DRQS-170879936

Signed-off-by: Nirbhay Choubey <nchoubey@bloomberg.net>