Fixing misuse of nonames in non-singlemeta code
The combination of non-singlemeta and nonames are problematic:
When nonames is on, we use the name "comdb2_meta" to open all meta's.
As a result, all tables end up sharing a common meta. The example
below demonstrates the problem:

```
$ cdb2sql -tabs nonsinglemeta local "exec procedure sys.cmd.send('stat compr')"
COMPRESSION FLAGS
These apply to new records only!
[sqlite_stat1    ] ODH: yes Compress: lz4      Blob compress: lz4       in-place updates: yes  instant schema change: yes
[sqlite_stat4    ] ODH: yes Compress: lz4      Blob compress: lz4       in-place updates: yes  instant schema change: yes
[t1              ] ODH: yes Compress: lz4      Blob compress: lz4       in-place updates: yes  instant schema change: yes

$ cdb2sql nonsinglemeta local "create table t2 (i int) OPTIONS rec none, blobfield zlib"

$ cdb2sql -tabs nonsinglemeta local "exec procedure sys.cmd.send('stat compr')"
COMPRESSION FLAGS
These apply to new records only!
[sqlite_stat1    ] ODH: yes Compress: none     Blob compress: zlib      in-place updates: yes  instant schema change: yes
[sqlite_stat4    ] ODH: yes Compress: none     Blob compress: zlib      in-place updates: yes  instant schema change: yes
[t1              ] ODH: yes Compress: none     Blob compress: zlib      in-place updates: yes  instant schema change: yes
[t2              ] ODH: yes Compress: none     Blob compress: zlib      in-place updates: yes  instant schema change: yes
```

The database has 3 tables which are all compressed using lz4.
A fourth table is added with no compression for data and zlib
for blob. All 4 tables now have the same compression settings.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>