{171755293}: Fixing race on clnt
After being freed by an appsock thread or a libevent event-loop, clnt
may still be read from another thread of control. This patch fixes it.

The original patch that attempted to address this issue, unfortunately,
triggered a different race condition between sql-thread-exiting and
newsql-cleanup. This patch adds a test to verify that both of these
2 races are addressed.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>