Do not donate socket if SSL_shutdown fails
SSL_shutdown() will fail if there is unread data on an SSL connection.
The patch changes the API to check the return code from SSL_shutdown(),
and to only donate the underlying socket after SSL_shutdown() succeeds.

While #3822 successfully reproduces a case in 7.0 that misleads the API
to incorrectly donate a undrained socket to sockpool, I've struggled
to get it to happen in 8.0. Therefore, in this patch I've also added
instrumentation to catch any protobuf response type other than LAST_ROW and
immediately before a RESET from sockpool, as we believe that such a packet
causes the behavior in the first place.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>