Repartitioning table
Currently changing partitioning on a table requires two schemachanges: the existing
partition needs to be dropped first, and only then a new partition can be created
on the table.

This patch extends alter-table to repartition a table within a single schemachange.
A new keyword 'REPARTITIONED' is introduced, to emphasize that this operation
moves data around and creates a new partition.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>