PRAGMA CHUNK
A chunk "transaction" requires to be wrapped in a transaction explicitly using BEGIN and COMMIT,
though it's not atomic. This usage has caused a lot of confusion. This patch proposes a new
syntax for chunk transactions, using SQLite's PRAGMA:

```
SET TRANSACTION CHUNK N
PRAGMA CHUNK=ON -- begin a chunk tran
INSERT INTO T VALUES(...)
PRAGMA CHUNK=OFF -- finish a chunk tran
```

The CHUNK pragma and `BEGIN..COMMIT` behave mostly identically, except that the CHUNK pragma
disables the skip-row optimization. To make these two work together, it would require the client
to recognize the pragma and treat it as a BEGIN, and that means client-side parsing and a new
client feature to be backwards compatible. Since chunk is mostly used in scheduled jobs that run
only periodically where performance isn't the greatest concern, I think I'm okay with this tradeoff.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>