{172658452}: Speeding up dohast bound parameter processing
Rebinding parameters may cause SQLite to recompile the statement in the hope
of a better query plan. A new AST is parsed out during the process, as well.

While it makes sense to recode the statement after parameters are changed,
it seems a bit wasteful to generate another identical AST. Hence in this patch,
the dohast struct is relocated from the parser to the sqlite3 db handle,
and is reused whenever possible.

In addition, a lookaside hash table for bound parameter names, is added in
this patch, turning an O(N^2) search into an O(1) hash table lookup. This change
greatly speeds up AST parsing for a large number of bound parameters.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>