{171428666}: column character encoding
A new `ENCODING` keyword is introduced in this patch, that specifies
a character set for a column. Invalid character encoding will be
rejected. So far only `"utf8"` and `NONE` are supported.

This allows users to conveniently create an indexable utf8 cstring column.
It's implemented as a check constraint using the `utf8_validate()` function.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>