Fix ar_compress and ar_decompress not set bug
The current behavior is that if lz4 exists on the default path (/opt/bb/bin/lz4), [[ ! -x $lz4 ]] would evaluate to false, and ar_compress and ar_decompress wouldn't be set to correct value and therefore undefined.

The expected behavior is that ar_compress and ar_decompress should be set in all circumstances.

Signed-off-by: Zixuan Zhang <zixuanzhang.x@gmail.com>