sp: Disable dohsql inside stored procedures
This fixes the systable_locking test. Failure is caused when clnt
callbacks are overwritten by running dohsql-able statement:
  db:exec("sql using dohsql")
  db:exec("any other sql stmt") -- fails

It is not sufficient to simply reset the callbacks at the end of the
statement. Interaction can get more involved. Procedure might be
traversing result of a dohsql'able statement, and for each row, might
want to run another sql statement. Or the other way around. Or two or
more dohsql-able statements. Feels like these callbacks should be set in
the Vdbe. For now, safer to just disable this if running in context of a
stored procedure.

Signed-off-by: Akshat Sikarwar <asikarwar1@bloomberg.net>