{174092326}: Fixing double-abort
#4143 attempts to fix double-abort by setting the berkdb txn struct
passed in by caller to NULL in `open_dbs()`, so that it'll become
no-op if caller aborts the transaction again. But instead of modifying
the actual DB_TXN struct, the PR ends up modifying the argument to
the function. This patch corrects this mistake.

The fastinit/truncate code is also modified to retry on deadlock from
llmeta. A test is added to reproduce the crash and to verifiy that
it works correctly with the bug fix.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>