{173871120}: Allowing cold-started replicants to participate in election
Currently replicants brought up afresh do not participate in election unless
they are marked 'caught up' after the global bdb_env is initialized. However
when half or more of the cluster are in this state, and the leader downgrades,
the cluster will then not be able to elect because of lack of quorum, as seen below

```
3333 2024/01/24 14:21:20 0x0x7fc1954f9700: calling for election with cluster of 3 nodes (3 connected) : c2btda-pw-053 c2btda-pw-928 c2btda-pw-922 ,  2.388000 secs timeout and priority 100
3343 2024/01/24 14:21:20 __rep_elect line 1453 not enough vote1s, failing
3344 2024/01/24 14:21:20 __rep_elect line 1533 ret is -30979
3345 2024/01/24 14:21:20 __rep_elect_done line 846 clearing PHASE1/PHASE2/TALLY, inelect is 1
3346 2024/01/24 14:21:20 __rep_elect_done called from __rep_elect line 1536
3347 2024/01/24 14:21:20 __rep_elect_done line 856 setting rep->egen from 168 to 169
3348 2024/01/24 14:21:20 __rep_elect line 1544 returning -30979 master .invalid egen is 168
3349 2024/01/24 14:21:20 failed to reach consensus in election with 2.388000 secs timeout. network timeout or received votes of a different egen
```

This patch gets rid of the 'caught-up' check. As long as a replicant is connected
to the leader, and the leader requests election, the replicant should always respond
to that request.

A new test electbug2 is added to reproduce this bug.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>