Use ReentrantLock instead of synchronized blocks
JDK 21 has introduced official support for virtual threads. It is a framework
for efficient processing of blocking tasks in the simple thread-per-request
style. Behind the scenes, the JVM runtime maintains a relatively small number
of platform threads. A virtual thread is mounted onto a platform thread for
execution. When the virtual thread blocks (e.g. on I/O) it is unmounted, and
the platform thread is then free to execute a different task. See
https://openjdk.org/jeps/444 for more details.

The current virtual threads implementation, however, does not play nicely with
`synchronized` blocks and methods, especially when blocking operations are
performed inside of them. In these situations the platform threads cannot be
freed, and the JEP recommends using `ReentrantLock` instead. This PR replaces
all usages of `synchronized` in `cdb2jdbc` with the recommended lock type.

Note:

https://github.com/bloomberg/comdb2/pull/4174 must be merged first to ensure
that the build of the library succeeds.

Signed-off-by: Viacheslav Chernyshev <v.chernyshev@outlook.com>