{174092326}: caching file versions in bdb_state struct
This patch caches file versions in the bdb_state struct, so that
we avoid reading from llmeta next time we need them. Note that
it does not fix places where wo do not unroll schemachange states
correctly on error from llmeta, but it does make it much less
likely for deadlocks to occur.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>