For multi-table schema change transactions, we need a LLMETA_SC_SEED per table, otherwise resume fails.  This is a follow-up on a previous resume fix.
NOTE: as it stands, code treats schema changes for which uuid=0 as standalone
and the rest as part of a multi-table ddl transaction.  If a master swings interrupts
the transaction, the resume is different in the two cases.
For standalone table schema changes, the replicant does not retry the bplog if it
detects a master swing, and returns an error. Instead, a new master resumes and
finalize the schema change.
In the second case, the master resumes each schema change but does not finalize it.
Instead, it is the replicant that retries the bplog, and the code on the master
detects the pending schema changes and only runs the finalize part once the
first part is completed.

NOTE2: as a consequence of the above, the single-node clusters do not finalize
a multi-table ddl txn upon restart. This will be the subject of a future PR.

Signed-off-by: Dorin Hogea <dhogea@bloomberg.net>