{172983130}: Limiting number of rows returned in comdb2_sc_history
It takes a whopping 75GiB memory to query the `comdb2_sc_history` table
of 2.4 million records, against the database in the linked ticket.

In this patch comdb2_sc_history and comdb2_sc_status are changed
to return the most recent 10K records.

A memory leak in the comdb2_sc_status system table is also fixed,
int this patch.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>