{174866577}: Fixing libevent stack overflow
The following pseudo code crashes the server, when sockpool is present

```
while true
  cdb2_open()
  /* do nothing */
  cdb2_close()
end
```

The server will call rd_hdr recursively, and eventually will overflow
its call stack, as shown below.

```
12 rd_hdr
13 rd_hdr
14 rd_hdr
...
32672 rd_hdr
32673 rd_hdr
32674 rd_hdr
32675 event_process_active_single_queue
32676 event_process_active
32677 event_base_loop
32678 net_dispatch
32679 start_thread
32680 clone
```

This patch fixes it: after multiple consecutive sql resets are received from
the same user, the user will be rescheduled back to the event queue, instead
of being processed inline.

Note that open-cdb2api is unaffected by this, as it always does a dbinfo
query which will break the server out of the recursion. On the other hand,
bb-cdb2api which has the skip-dbinfo optimization is prone to this bug.
I have added an internal `donate_ddos` test to demonstrate the bug.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>