{174899913}: Fixing sporadic ENOENT errors
`comdb2_get_tmp_dir()` isn't thread-safe and may return an empty string.
This causes the sqlite sorter to fail with ENOENT intermittently,
when opening a temporary file.

This patch correctly sets the `once` flag in `comdb2_get_tmp_dir()`,
to make it thread-safe for calls after the very first call.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>