Tweak transchunk_manyins to workaround chunk-transaction bug
Send fewer queries, but do same number of inserts with same chunk size.

This is happening even in 7.0, and in 8.0 (with and without libevent.)
The issue is that client times out sending query to server (primary
bug), reconnects, starts sending further queries. Secondary bug is that
the new connection is not SET with chunk, but detects that it is in
transaction. So client doesn't send new cnonce for every statement.
Server uses cnonce since not in chunk anymore, and fails on second query
on new connection, as it detects replay on the cnonce.

Signed-off-by: Akshat Sikarwar <asikarwar1@bloomberg.net>