Tweaking the current_connections metric
This patch excludes connections owned by either sockpool or the in-process cache from
the current_connections metric. This is one of the metrics that we use internally to
monitor database capacity. Since it is very little overhead for libevent based network
to maintain idle connections, it makes sense to exclude pooled connections from the metric.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>