Address bug
in a txn
(1) constraint is added
(2) constraint is satisfied with an alter
during the scdone step for (2), replicants will
assume that, since there is a constraint pointing to
the table getting altered, the target key
must have existed in the table before it got altered.

This was previously a sound assumption because it was not possible
to first add a constraint before satisfying it in the same transaction,
which is the only case where this can validly occur.

The changes in this commit 'pass' the compatibility check if
we can't find the key that we're trying to validate in the
pre-alter dbtable.

TODO: make sure other compatibility checks in ondisk_schema_change aren't buggy.

Signed-off-by: Morgan Douglas <mdouglas47@bloomberg.net>