Distributed table creation and drop for generic sharding.
Introduces syntax:
CREATE TABLE .... PARTITIONED BY TESTGENSHARD(dbname1, dbname2, dbname3, dbname4)

Dropping a partitioned table using the partition name drops the whole partition in one transaction.

Use a stub to interface with rest of sharding code, to provide shardnames, number of shards (hardcded 4), and location of databases.
The only flexible parameter is for the database names used to store the shards.
The stub code is marked to the best ability by the framing message:
"THIS SECTION IS A STUB; TO BE REPLACED BY ACTUAL PARTITION SCHEMA".
"END: THIS SECTION IS A STUB; TO BE REPLACED BY ACTUAL PARTITION SCHEMA".

In that sense, the code does not store anything in llmeta, and passes all configuration through SET and schema change options.
It simplifies the partition by only using an table alias instead of a full-fledged view.

Testcase will create the partition, check it and remove it.
[The row routing for read and writes is part of a separate PR, so no actual rows are read/written to the partition].

Signed-off-by: Dorin Hogea <dhogea@bloomberg.net>