Dedicated queue writers
Increasing the number of writers on the master does not always improve
database performance. More often than not it increases contention
and hurts performance. But queues may be an exception: only one client
can consume from a queue at a time, that it could make sense for
queue consumers to run in parellel as much as possible.

On a tunable, this patch dispatches queue consumes to dedicated
writer threads. The handle_buf code is also refactored to use thdpool.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>