{178644353}: Increasing temptable bufferpool size
Currently temptable bufferpool size is configured to 512KB.
We bump it by 25% and bring it to around 640KB, which is 10 pages.

10 pages are enough to split a leaf page in a btree of height 4,
or remove an empty leaf page in a btree of height 7.

The diagram below demonstrates the split case, which requires 10 pages
to split all the way up to the root, in a btree of heigh 4.

```
meta-page                                  1 page

level 4:    root                           1 pages
             \
level 3:      internal - new-internal      2 pages
               \
level 2:        internal - new-internal    2 pages
                 \
level 1:  prev -  split - split - next     4 pages

```

The diagram below demonstrates the relink case, which requires 10 pages
to empty a page, all the way up to the root, in a btree of height 7.

```
meta-page                                  1 page

level 7:    root                           1 page
             \
level 6:      internal                     1 page
               \
level 5:        internal                   1 page
                 \
level 4:          internal                 1 page
                   \
level 3:            internal               1 page
                     \
level 2:              internal             1 page
                       \
level 1:         prev - deleted - next     3 page
```

Normally the btree height won't be very large, because the key size is limited
to 512 bytes there. However, temptables are not subject to the 512-byte limit,
and large keys close to the overflow threshold (eg, just a tad under 16K) can
create a very tall btree, and exhaust the bufferpool.

This patch bumps the temptable buffer size to 1MB, which brings the max number
of temptable pages to 19, which is enough to split a btree of height 7, or to
empty a page in a btree of height 16. We would run out of disk space first before
reaching such a height.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>