Revert "A hint query that has NO full statement and is part of a transactio  is NOT retryable."
This reverts commit c9c76eeb8b4382059ee0f53b63b8cf177052f745.

This commit ends up breaking sql-hint: if a query isn't found for a hint string,
server needs to return a retryable error so that client can re-send the query in its full form.
It should not be treated as a fatal error, even if the client has begun a transaction.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>