Allowing omitting the "T" separator from a datetime string
Our customers have issue using spark to create a dataframe on a
datetime column, because spark interpolates the provided datetime
string, but drops the "T" separator when filling the range.

It seems that omitting the "T" character is fairly acceptable, and
ISO 8601 does allow it to be omitted in a few cases. This is a
very simple patch to allow it.

Signed-off-by: Rivers Zhang <hzhang320@bloomberg.net>