ci: do not run workflow on pushes to `main`
In the event that a developer is using a personal access token that does not have `workflow` permissions, pushing to `main` on the developer's fork will fail. Since we run the workflow on PRs, additionally checking commits to `main` seems redundant anyway. Do not run checks on pushes to main to avoid disrupting developers.