`{180685083}`: Do not block statreq updates during schema change
It seems this logic has existed since before schema lock was a thing. The
critical section of `statthd` that inspects `dbenv` now acquires the schema
lock, so this check should no longer be necessary. Remove the check in `statthd`
for whether schema change is in progress so that statreqs are not missing on the
master during schema change.

Signed-off-by: Shoumyo Chakravorti <schakravorti@bloomberg.net>