{178862843} fix(crc32c): avoid misaligned memory accesses
This bug was initially reported by @hallfox (thank you!). The issue is that
`crc32c_sse` may generate misaligned memory accesses because of the order of
operations it performs:

1. Compute the checksum until an 8 byte-aligned address is reached.
2. Compute the checksum until an exact multiple of 1024 bytes is remaining.
3. Compute the checksum for the remaining bytes in chunks of 1024 bytes.

Unfortunately, Step 2 can kick us out of alignment, which breaks the
requirement that we be at an alignment boundary at the beginning of Step 3.

We can ensure alignment by doing the following instead:

1. Compute the checksum until an 8 byte-aligned address is reached.
2. Compute the checksum for as many 1024 byte chunks as possible.
3. Compute the checksum for the remaining bytes.

Reproducing the original issue uncovered another one, which was that some
left-shifts performed in `crc32c_software` could result in overflow since we
did not explicitly upcast the shifted bytes (`uint8_t`) to the target storage
type (`uint32_t`). This is fixed by upcasting before shifting.

Signed-off-by: Shoumyo Chakravorti <schakravorti@bloomberg.net>