{180685083} feat(reqlog): post schema change activity
Statreqs tell us what the database is busy doing. If any schema changes are
running, that is significant and worth reporting in statreqs.

Schema change stats are logged in a dedicated section that looks like this:

```text
11/29 02:55:00:   SCHEMA CHANGE STATS
11/29 02:55:00:       table 't' records converted 269805 diff 169469 rate 84734 r/s adds 0 updates 0 deletes 77
```

These stats will only be logged on the node processing the schema change, i.e.
the master. Updated the `reqlog_update_during_sc` test case to ensure that these
stats are generated on master during schema change. Additionally, updated the
test case to be less flaky.

Signed-off-by: Shoumyo Chakravorti <schakravorti@bloomberg.net>