sqlwriter: Defer creating eventbase
Every eventbase creates a socketpair for internal use by libevent. The
eventbase is used only when SQL request is running. Stop creating these
when connection is accepted. Instead, create a cache of eventbases, so
these can be reused. Grab an eventbase when SQL request is dispatched.
This way we should only have (#sql-thd + #sql-queue) number of
eventbases, and reduces how many file-descriptors are consumed when a
connection is accepted.

Signed-off-by: Akshat Sikarwar <asikarwar1@bloomberg.net>